+++ yampt.translator/source/dialog/settings/translator_settings_dialog.hpp
@@ -10,6 +10,7 @@ class QStackedWidget;
 class settings_store_t;
 class appearance_settings_view_t;
 class language_settings_view_t;
+class translation_settings_view_t;
 class shortcuts_settings_view_t;
 
 class translator_settings_dialog_t : public QDialog
@@ -38,5 +39,6 @@ private:
 
 	appearance_settings_view_t * m_appearance_view = nullptr;
 	language_settings_view_t * m_language_view = nullptr;
+	translation_settings_view_t * m_translation_view = nullptr;
 	shortcuts_settings_view_t * m_shortcuts_view = nullptr;
 };