+++ yampt.translator/source/model/yaml_document.hpp
@@ -28,6 +28,9 @@ public:
 	int translated_count() const override;
 	int total_count() const override;
 
+	std::set<rec_type_t> supported_types() const override;
+	std::set<status_t> supported_statuses() const override;
+
 	void set_dirty(bool dirty) override;
 
 	const std::string & foreign_path() const;